+++ packages/debug-test/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - debug-test
 
-This log was last generated on Sat, 12 Nov 2022 06:17:23 GMT and should not be manually modified.
+This log was last generated on Tue, 15 Nov 2022 06:16:38 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.0.0
+
+Tue, 15 Nov 2022 06:16:38 GMT
+
+### Patches
+
+- Bump @react-native-windows/automation to v0.3.3
 
 ## 0.0.0